pyproject.toml
[project]
name = "langchain-deepdive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.12.1",
    "grandalf>=0.8",
    "langchain>=0.3.24",
    "langchain-google-genai>=2.1.3",
    "langfuse>=2.60.3",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
